+++ tests/test_version.py
@@ -2,4 +2,4 @@ import geoworkbench
 
 
 def test_package_version() -> None:
-    assert geoworkbench.__version__ == "0.7.18"
+    assert geoworkbench.__version__ == "0.7.19"